Working on positioning currentNewPost element
I started with a rather complicated structure but finished with a far
less complicated one. Ultimately, I realized that all I needed to do was
give the element that I want to move a relative position and move it
accordingly.

As a result of the need to simply move the element relative to its
original position, I am going to implement a function in the
SlideshowCtrl to dynamically adjust the positioning of the element
depending on the post's height. The goal is to have the bottom edge of
the photo constantly line up with the bottom edge of the smallFilmStrip.