Setting max width in angular capsule view
We needed to make sure that the spacing between pictures in the capsule
doesn't become so wide that it turns the capsule into something really
ugly. The easiest solution was simply to set a max-width and to add a
"row" class to the image container so that it was centered.