Making fixes to discounts
After Lindsey tried to create a new discount and use it, there was a
number of issues that she had that immediately became evident. There was
no feedback on the discount creation form to force you to enter the
correct information. Hence, I needed to correct that. In addition, there
needed to be a discount show page so you don't have to go to the edit
view to see the discount specs and potentially change them.

In addition, I needed to make some changes to the discount controller
because the Date parsing caused an error when render the new action with
errors.