Small changes to first iteration of slideshow
I made it so the top bar is hidden when in the slideshow view and added
a link to go back to the capsule.

In addition, I added a couple of different versions of the slideshow in
order to illustrate new photos that are coming in. We were not entirely
happy with the recent rendition. Hence, I set it up so there are a
couple of different styles so we can ask people's opinions and try to
figure out which is the best.