Adding the ability to inspect the notification count in EmailEngine
I realized that the EmailEngine class variable notification_count might
be backing up and day after day taking up more memory. Hence, I would
like to be able to look at it and see whether it is build a giant hash
or whether it gets cleared out every time the server is reloaded.