Adding a method for getting the body text with multipart
When sending a text message with only text from a Sprint carrier phone,
we ran into issues. In the end, it appears as though they send their
messages as multipart rather than simply putting the text in the body of
the main message. Hence, I needed to access that part and get the text
from that.