Fixing the page load bug.
This was in fact caused by a recursive function that I was using to
build the images in the CoverPhotoCtrl. Without realizing it, I was
making state changes outside the scope of the function. This is once
again, the pain in the ass about working with a seemingly functional
languages that is not actually immutable. I'm used to thinking of it in
a manner where I'm not changing the state of anything outside a
function. Unfortunately, that is not the way that javascript works.